+++ src/components/MovieTile/MovieTags/index.js
@@ -1,0 +1,9 @@
+import { MovieTileTags, MovieTileTag } from "./styled";
+
+export const MovieTags = ({ tags }) => (
+  <MovieTileTags>
+    {tags
+      ? tags.map((tag) => <MovieTileTag key={tag}>{tag}</MovieTileTag>)
+      : ""}
+  </MovieTileTags>
+);
