+++ src/components/MovieTile/styled.js
@@ -1,5 +1,4 @@
 import styled from "styled-components";
-import { ReactComponent as StarIcon } from "../../images/VectorStar.svg";
 
 export const MovieTileContent = styled.li`
   padding: 16px;
@@ -73,56 +72,4 @@ export const MovieTileSubtitle = styled.p`
   @media (max-width: ${({ theme }) => theme.breakpoint.mobile}px) {
     font-size: 12px;
   }
-`;
-
-export const MovieTileTags = styled.ul`
-  margin: 0;
-  list-style: none;
-  padding-left: 0;
-  display: flex;
-  flex-wrap: wrap;
-  gap: 8px;
-`;
-
-export const MovieTileTag = styled.li`
-  padding: 8px 16px;
-  border-radius: 4px;
-  font-size: 14px;
-  line-height: 1.4;
-  background-color: ${({ theme }) => theme.color.tagBG};
-
-  @media (max-width: ${({ theme }) => theme.breakpoint.mobile}px) {
-    padding: 4px 8px;
-    font-size: 10px;
-  }
-`;
-
-export const MovieTileRating = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 12px;
-  font-size: 16px;
-  font-weight: ${({ theme }) => theme.fontWeight.semiBold};
-  line-height: 1.5;
-
-  @media (max-width: ${({ theme }) => theme.breakpoint.mobile}px) {
-    margin: 0;
-    gap: 8px;
-    font-size: 12px;
-    line-height: 1.3;
-  }
-`;
-
-export const StyledStarIcon = styled(StarIcon)`
-  width: ${({ big }) => (big ? "40" : "24") + "px"};
-  transition: 0.6s;
-
-  @media (max-width: ${({ theme }) => theme.breakpoint.mobile}px) {
-    width: ${({ big }) => (big ? "24" : "16") + "px"};
-  }
-`;
-
-export const MovieTileRatingVotes = styled.span`
-  font-weight: ${({ theme }) => theme.fontWeight.regular};
-  color: ${({ theme }) => theme.color.darkerGrey};
 `;
